gcs: Rename package for The Things Kickstarter Gateway
