+++ cmd/internal/shared/gatewayconfigurationserver/config.go
@@ -26,8 +26,8 @@ var DefaultGatewayConfigurationServerConfig = gatewayconfigurationserver.Config{
 }
 
 func init() {
-	DefaultGatewayConfigurationServerConfig.TheThingsGateway.Default.UpdateChannel = "stable"
-	DefaultGatewayConfigurationServerConfig.TheThingsGateway.Default.MQTTServer = "mqtts://" + gs.DefaultGatewayServerConfig.MQTTV2.PublicTLSAddress
-	DefaultGatewayConfigurationServerConfig.TheThingsGateway.Default.FirmwareURL = "https://ttkg-fw.thethingsindustries.com/v1"
+	DefaultGatewayConfigurationServerConfig.TheThingsKickstarterGateway.Default.UpdateChannel = "stable"
+	DefaultGatewayConfigurationServerConfig.TheThingsKickstarterGateway.Default.MQTTServer = "mqtts://" + gs.DefaultGatewayServerConfig.MQTTV2.PublicTLSAddress
+	DefaultGatewayConfigurationServerConfig.TheThingsKickstarterGateway.Default.FirmwareURL = "https://ttkg-fw.thethingsindustries.com/v1"
 	DefaultGatewayConfigurationServerConfig.BasicStation.Default.LNSURI = "wss://" + shared.DefaultPublicHost + gs.DefaultGatewayServerConfig.BasicStation.ListenTLS
 }
