+++ pkg/gatewayconfigurationserver/ttkg/http_gateway.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package gatewayconfigurationserver
+package ttkg
 
 import (
 	"fmt"
